+++ Marked/Src/cutscene9.c
@@ -21,32 +21,32 @@ void Cutscene9_Update() {
 	CP_Settings_Fill(CP_Color_Create(255, 255, 255, 255));
 	time += CP_System_GetDt();
 	if (time < .5)
-		CP_Font_DrawText("YOU", windowWidth / 2, windowHeight / 2);
+		CP_Font_DrawText("YOU", (float)windowWidth / 2.f, (float)windowHeight / 2.f);
 	else if (time < 1)
-		CP_Font_DrawText("THINK", windowWidth / 2, windowHeight / 2);
+		CP_Font_DrawText("THINK", (float)windowWidth / 2.f, (float)windowHeight / 2.f);
 	else if (time < 1.5)
-		CP_Font_DrawText("YOU'RE", windowWidth / 2, windowHeight / 2);
+		CP_Font_DrawText("YOU'RE", (float)windowWidth / 2.f, (float)windowHeight / 2.f);
 	else if (time < 2) {
 		CP_Settings_Fill(CP_Color_Create(128, 0, 0, 255));
-		CP_Font_DrawText("FAST?", windowWidth / 2, windowHeight / 2);
+		CP_Font_DrawText("FAST?", (float)windowWidth / 2.f, (float)windowHeight / 2.f);
 	}
 	else if (time < 3)
 		CP_Settings_Fill(CP_Color_Create(255, 255, 255, 255));
 	else if (time < 3.5)
-		CP_Font_DrawText("WELL...", windowWidth / 2, windowHeight / 2);
+		CP_Font_DrawText("WELL...", (float)windowWidth / 2.f, (float)windowHeight/ 2.f);
 	else if (time < 3.8)
-		CP_Font_DrawText(".", windowWidth / 2, windowHeight / 2);
+		CP_Font_DrawText(".", (float)windowWidth / 2.f, (float)windowHeight / 2.f);
 	else if (time < 4.2)
-		CP_Font_DrawText("..", windowWidth / 2, windowHeight / 2);
+		CP_Font_DrawText("..", (float)windowWidth / 2.f, (float)windowHeight / 2.f);
 	else if (time < 4.7)
-		CP_Font_DrawText("...", windowWidth / 2, windowHeight / 2);
+		CP_Font_DrawText("...", (float)windowWidth / 2.f, (float)windowHeight / 2.f);
 	else if (time < 5) {
 		CP_Settings_Fill(CP_Color_Create(128, 0, 0, 255));
-		CP_Font_DrawText("THINK", windowWidth / 2, windowHeight / 2);
+		CP_Font_DrawText("THINK", (float)windowWidth / 2.f, (float)windowHeight / 2.f);
 	}
 	else if (time < 5.3) {
 		CP_Settings_Fill(CP_Color_Create(128, 0, 0, 255));
-		CP_Font_DrawText("FASTER", windowWidth / 2, windowHeight / 2);
+		CP_Font_DrawText("FASTER", (float)windowWidth / 2.f, (float)windowHeight / 2.f);
 	}
 	else if (time > 5.3)
 		TransitScene(Level9);
